Put guard rails in place to protect against crashes due to bad input. TODO fix Weapon Skill connection on Weapons.

+++ Master Sheet/src/main/java/com/example/app/ViewModel/InventoryDetailsViewModel.java
@@ -1,7 +1,9 @@
 package com.example.app.ViewModel;
 
 import android.app.Application;
+import android.content.Context;
 import android.util.Log;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.lifecycle.AndroidViewModel;
@@ -24,8 +26,10 @@ public class InventoryDetailsViewModel extends AndroidViewModel {
         repository = AppRepository.getInstance(getApplication());
     }
 
-    public void deleteSkill() {
-        repository.deleteInventory(item.getValue());
+    public void deleteSkill(int id) {
+        if(id !=-1) {
+            repository.deleteInventory(item.getValue());
+        }
     }
 
     public void saveItem(final Inventory newItem) {
